Moved Window and Gui to voxen::client namespace
