+++ src/CMakeLists.txt
@@ -14,13 +14,13 @@ set (VOXEN_SOURCES
 	src/client/vulkan/texture_manager.cpp
 	src/client/vulkan/vulkan_render.cpp
 	src/client/debug_draw_octree.cpp
+	src/client/gui.cpp
 	src/client/input_manager.cpp
 	src/client/render.cpp
 	src/client/window.cpp
 	src/common/player.cpp
 	src/common/terrain.cpp
 	src/common/world.cpp
-	src/common/gui.cpp
 	src/common/gameview.cpp
 	src/common/config.cpp
 	src/util/assert.cpp
